+++ resources/js/common.js
@@ -3,3 +3,5 @@ import '@fortawesome/fontawesome-free/js/all.min.js';
 import '../sass/bulma-theming.scss';
 import '../css/bulma-override.css';
 import '../css/common.css';
+import '../css/footer.css';
+import '../css/flag-icon.min.css';