+++ resources/views/footer.blade.php
@@ -1,0 +1,11 @@
+<footer class="container is-fluid">
+    <div class="columns">
+        <div class="column is-12">
+            <hr id="footer-hr">
+            <a id="french-link" href="{{ url('/lang/fr') }}"><i class="flag-icon flag-icon-fr"></i></a>
+            <a id="english-link" href="{{ url('/lang/en') }}"><i class="flag-icon flag-icon-gb"></i></a>
+            <a id="github-link" href="https://www.github.com/Lenophie/MoniCAG/"> {{__('messages.titles.github')}} <i class="fab fa-github"></i></a>
+        </div>
+    </div>
+</footer>
+