+++ resources/views/template.blade.php
@@ -12,6 +12,7 @@
 
     <body>
         @yield('content')
+        @include('footer')
     </body>
 
     <script type="text/javascript" src="{{ asset('js/common.js') }}"></script>