+++ changelog.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 * Testing
     * Locale setting unit test
+    
+### Changed
+
+* Views
+    * Made home footer a common footer for every page
 
 ## [0.15.0] - 2018-10-20